Correct data type

+++ src/mach3sbitools/diagnostics/sbc.py
@@ -43,7 +43,9 @@ class SBCDiagnostic:
         self.prior_predictives: torch.Tensor | None = None
 
     def create_prior_samples(self, num_prior_samples: int) -> None:
-        self.prior_samples = self.inference_handler.prior.sample((num_prior_samples,))
+        self.prior_samples = self.inference_handler.prior.sample(
+            (num_prior_samples,)
+        ).to(float)
 
         prior_predictives_np = np.array(
             [
@@ -54,7 +56,9 @@ class SBCDiagnostic:
             ],
             dtype=np.double,
         )
-        self.prior_predictives = self._device_handler.to_tensor(prior_predictives_np)
+        self.prior_predictives = self._device_handler.to_tensor(
+            prior_predictives_np
+        ).to(float)
 
     def _check_prior_sampled(self):
         if self.prior_predictives is None:
